core: move chain dump test into its own package
